+++ CHANGELOG.md
@@ -1,8 +1,12 @@
+# Future
+
+- Changed: `requireMatchingFilename` is now more configurable.
+
 # Release 0.0.2 (2014-09-06)
 
-- Added: requireMatchingFilename rule.
+- Added: `requireMatchingFilename` rule.
 - Validate option values for rules.
 
 # Release 0.0.1 (2014-09-03)
 
-- Initial release with requireAngularDependencyOrder rule.
+- Initial release with `requireAngularDependencyOrder` rule.